Extract CacheIndexStatement to reduce duplicate logic for MySQL and doris
