+++ parser/sql/dialect/doris/src/main/java/org/apache/shardingsphere/sql/parser/doris/visitor/statement/type/DorisDALStatementVisitor.java
@@ -206,12 +206,12 @@ import org.apache.shardingsphere.sql.parser.sql.dialect.statement.doris.dal.Dori
 import org.apache.shardingsphere.sql.parser.sql.dialect.statement.doris.dal.DorisUninstallComponentStatement;
 import org.apache.shardingsphere.sql.parser.sql.dialect.statement.doris.dal.DorisUninstallPluginStatement;
 import org.apache.shardingsphere.sql.parser.sql.dialect.statement.doris.dal.DorisUseStatement;
-import org.apache.shardingsphere.sql.parser.sql.dialect.statement.doris.segment.CacheTableIndexSegment;
+import org.apache.shardingsphere.sql.parser.sql.common.segment.dal.CacheTableIndexSegment;
 import org.apache.shardingsphere.sql.parser.sql.dialect.statement.doris.segment.CloneActionSegment;
 import org.apache.shardingsphere.sql.parser.sql.dialect.statement.doris.segment.CloneInstanceSegment;
 import org.apache.shardingsphere.sql.parser.sql.dialect.statement.doris.segment.LoadTableIndexSegment;
-import org.apache.shardingsphere.sql.parser.sql.dialect.statement.doris.segment.PartitionDefinitionSegment;
-import org.apache.shardingsphere.sql.parser.sql.dialect.statement.doris.segment.PartitionSegment;
+import org.apache.shardingsphere.sql.parser.sql.common.segment.dal.PartitionDefinitionSegment;
+import org.apache.shardingsphere.sql.parser.sql.common.segment.dal.PartitionSegment;
 import org.apache.shardingsphere.sql.parser.sql.dialect.statement.doris.segment.ResetMasterOptionSegment;
 import org.apache.shardingsphere.sql.parser.sql.dialect.statement.doris.segment.ResetOptionSegment;
 import org.apache.shardingsphere.sql.parser.sql.dialect.statement.doris.segment.ResetSlaveOptionSegment;
