+++ parser/sql/statement/src/main/java/org/apache/shardingsphere/sql/parser/sql/common/segment/dal/CacheTableIndexSegment.java
@@ -15,25 +15,25 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.sql.dialect.statement.mysql.segment;
+package org.apache.shardingsphere.sql.parser.sql.common.segment.dal;
 
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-import org.apache.shardingsphere.sql.parser.sql.common.segment.SQLSegment;
-import org.apache.shardingsphere.sql.parser.sql.common.value.identifier.IdentifierValue;
+import org.apache.shardingsphere.sql.parser.sql.common.segment.ddl.index.IndexSegment;
+import org.apache.shardingsphere.sql.parser.sql.common.segment.generic.table.SimpleTableSegment;
+import org.apache.shardingsphere.sql.parser.sql.dialect.statement.doris.segment.TableIndexSegment;
+
+import java.util.Collection;
+import java.util.LinkedList;
 
 /**
- * Partition segment.
+ * Cache table index segment.
  */
-@RequiredArgsConstructor
 @Getter
-@Setter
-public final class PartitionSegment implements SQLSegment {
+public final class CacheTableIndexSegment extends TableIndexSegment {
     
-    private final int startIndex;
+    private final Collection<IndexSegment> indexes = new LinkedList<>();
     
-    private final int stopIndex;
-    
-    private final IdentifierValue name;
+    public CacheTableIndexSegment(final int startIndex, final int stopIndex, final SimpleTableSegment table) {
+        super(startIndex, stopIndex, table);
+    }
 }
